Report or give feedback to the GUI client about running test units

These functions are usually not called from the command line. They return data to compatible GUI clients, like Komodo Edit with the SciViews-K extension.

guiTestReport(object, sep = "\t", path = NULL, ...)
guiTestFeedback(object, path = NULL, ...)
guiSuiteList(sep = "\t", path = NULL, compare = TRUE)
guiSuiteAutoList(...)

Arguments

object
a svUnitData object.

...
not used currently.

sep
field separator to use in the results.

path
path where to write a 'Suites.txt' file with the list of currently available test suites (to be used by the GUI client). If NULL, no file is written (by default).

compare
do we compare the list of available test suite and return something to the GUI client only if there are changes in the list? This is used (when TRUE) to avoid unnecessary multiple processings of the same list by the GUI client.

Value

guiSuiteList() returns the list of available test suites invisibly. guiSuiteAutoList() is used to establish a callback to automatically list the available test suites in the GUI. It is not intended to be called directly by the user. The other functions just return TRUE invisibly. They are used for their side effect of sending data to compatible GUI clients.

See also