Objects of class 'svSuiteData' contain results from running test suites

The 'svSuiteData' object contains results of all test run in one or more test suites. The checkxxx() functions and the runTest() method generate data (objects 'svTestData') contained in the default 'svSuiteData' named .Log and located in .GlobalEnv. It is then possible to display and report information it contains in various ways to analyze the results.

is.svSuiteData(x)

# S3 method for svSuiteData
stats(object, ...)

metadata(object, ...)

# S3 method for svSuiteData
metadata(
  object,
  fields = c("R.version", "sessionInfo", "time", "description"),
  ...
)

# S3 method for svSuiteData
print(x, all = FALSE, file = "", append = FALSE, ...)

# S3 method for svSuiteData
summary(object, ...)

protocol(object, type = "text", file = "", append = FALSE, ...)

# S3 method for default
protocol(object, type = "text", file = "", append = FALSE, ...)

# S3 method for svSuiteData
protocol(object, type = "text", file = "", append = FALSE, ...)

protocol_text(object, file = "", append = FALSE, ...)

# S3 method for svSuiteData
protocol_text(object, file = "", append = FALSE, ...)

protocol_junit(object, ...)

# S3 method for svSuiteData
protocol_junit(object, file = "", append = FALSE, ...)

Arguments

x
Any kind of object, or a 'svSuiteData' object in the case of print().

object
A 'svSuiteData' object.

...
Further arguments to pass to methods. Not used yet.

fields
Character vector. The name of all metadata items you want to extract for the object. The default value is an exhaustive list of all available metadata (i.e., defined by default) in the object, but you can add more: just add a corresponding attribute to your object.

all
Do we print concise report for all test, or only for the tests that fail or produce an error?

file
Character. The path to the file where to write the report. If file = "", the protocol report is output to the console.

append
Do we append to this file?

type
Character. The type of protocol report to create. For the moment, only type = "text" and type = "junit" are supported, but further types (HTML, LaTeX, Wiki, etc.) could be provided later.

Value

is.svSuiteData() returns TRUE if the object is an 'svSuiteData'. The various methods serve to extract or print content in the object.

Details

A 'svSuiteData' is, indeed, an environment. The results for the various tests runs are in non hidden (i.e., names not starting with a dot) objects that are of class 'svTestData' in this environment. Various other objects that control the execution of the test, their context, etc. are contained as hidden objects with name starting with a dot. Note that using an environment instead of a list for this object allows for a call by reference instead of a usual call by value in R, when passing this object to a function. This property is largely exploited in all svUnit functions to make sure results of test runs are centralized in the same log ('svSuiteData' object).

See also

Examples

clearLog()    # Clear any existing log

# Run some tests
checkTrue(1 < 2)
checkException(log("a"))
foo <- function(x, y = 2)
  return(x * y)
test(foo) <- function() {
  checkEqualsNumeric(4, foo(2))
  checkEqualsNumeric(6, foo(2, nonexisting))
  checkTrue(is.test(foo))
  warning("This is a warning")
  cat("Youhou from test!\n")  # Don't use, except for debugging!
  checkTrue(is.test(test(foo)))
  checkIdentical(attr(foo, "test"), test(foo))
  checkException(foo(2, nonexisting))
  #DEACTIVATED("My deactivation message")
  checkException(foo(2))  # This test fails
}
runTest(foo)

#> * : checkEqualsNumeric(6, foo(2, nonexisting)) run in less than 0.001 sec ... **ERROR**
#> Error in foo(2, nonexisting) : object 'nonexisting' not found
#> Youhou from test!
#> * : checkException(foo(2)) run in less than 0.001 sec ... **FAILS**
#> No exception generated!

# Now inspect the log, which is a 'svSuiteData' object
is.svSuiteData(Log())

#> [1] TRUE

stats(Log())

#>                kind timing                time unit msg
#> test(foo) **ERROR**  0.001 2020-04-13 16:05:08         
#> eval             OK  0.000 2020-04-13 16:05:08         

metadata(Log())

#> $.R.version
#>                _                           
#> platform       x86_64-apple-darwin15.6.0   
#> arch           x86_64                      
#> os             darwin15.6.0                
#> system         x86_64, darwin15.6.0        
#> status                                     
#> major          3                           
#> minor          6.3                         
#> year           2020                        
#> month          02                          
#> day            29                          
#> svn rev        77875                       
#> language       R                           
#> version.string R version 3.6.3 (2020-02-29)
#> nickname       Holding the Windsock        
#> 
#> $.sessionInfo
#> R version 3.6.3 (2020-02-29)
#> Platform: x86_64-apple-darwin15.6.0 (64-bit)
#> Running under: macOS Catalina 10.15.4
#> 
#> Matrix products: default
#> BLAS:   /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRblas.0.dylib
#> LAPACK: /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRlapack.dylib
#> 
#> locale:
#> [1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] svUnit_1.0.0
#> 
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_1.0.4.6       rstudioapi_0.11    xml2_1.3.1         whisker_0.4       
#>  [5] knitr_1.28         rematch2_2.1.1     magrittr_1.5       MASS_7.3-51.5     
#>  [9] R6_2.4.1           rlang_0.4.5        fansi_0.4.1        httr_1.4.1        
#> [13] tools_3.6.3        xfun_0.13          cli_2.0.2          withr_2.1.2       
#> [17] htmltools_0.4.0    ellipsis_0.3.0     yaml_2.2.1         assertthat_0.2.1  
#> [21] digest_0.6.25      rprojroot_1.3-2    tibble_3.0.0       lifecycle_0.2.0   
#> [25] pkgdown_1.5.1.9000 crayon_1.3.4       processx_3.4.2     purrr_0.3.3       
#> [29] vctrs_0.2.4        fs_1.4.1           ps_1.3.2           curl_4.3          
#> [33] evaluate_0.14      memoise_1.1.0      glue_1.4.0         rmarkdown_2.1     
#> [37] compiler_3.6.3     pillar_1.4.3       desc_1.2.0         backports_1.1.6   
#> [41] highlight_0.5.0    pkgconfig_2.0.3   
#> 
#> $.time
#> [1] "2020-04-13 16:05:08 UTC"
#> 

Log()    # Print method

#> = A svUnit test suite run in less than 0.1 sec with:
#> 
#> * test(foo) ... **ERROR**
#> * eval ... OK
#> 
#> 
#> == test(foo) run in less than 0.1 sec: **ERROR**
#> 
#> //Pass: 5 Fail: 1 Errors: 1//
#> 
#> * : checkEqualsNumeric(6, foo(2, nonexisting)) run in less than 0.001 sec ... **ERROR**
#> Error in foo(2, nonexisting) : object 'nonexisting' not found
#> 
#> * : checkException(foo(2)) run in less than 0.001 sec ... **FAILS**
#> No exception generated!
#> 
#> == eval run in less than 0.1 sec: OK
#> 
#> //Pass: 2 Fail: 0 Errors: 0//
#> 

summary(Log())

#> = A svUnit test suite run in less than 0.1 sec with:
#> 
#> * test(foo) ... **ERROR**
#> * eval ... OK
#> 
#> 
#> == test(foo) run in less than 0.1 sec: **ERROR**
#> 
#> //Pass: 5 Fail: 1 Errors: 1//
#> 
#> === Failures
#> [7] : checkException(foo(2))
#> 
#> === Errors
#> [2] : checkEqualsNumeric(6, foo(2, nonexisting))
#> 
#> 
#> == eval run in less than 0.1 sec: OK
#> 
#> //Pass: 2 Fail: 0 Errors: 0//
#> 

if (FALSE) {
# To get a print of the test protocol on file, use:
protocol(Log(), type = "text", file = "RprofProtocol.out")
file.show("RprofProtocol.out")
unlink("RprofProtocol.out")
}

rm(foo)

if (FALSE) {
# Profiling of very simple test runs
library(utils)
createLog(description = "test profiling", deleteExisting = TRUE)
imax <- 3
jmax <- 100
l <- 50
Rprof()
for (i in 1:imax) {
  # Change the context for these tests
  .Log$..Test <- paste("Test", i, sep = "")
  .Log$..Tag <- paste("#", i, sep = "")
  res <- system.time({
    for (j in 1:jmax) checkTrue(i <= j, "My test")
  }, gcFirst = TRUE)[3]
  print(res)
  flush.console()
}
Rprof(NULL)
# Look at profile
summaryRprof()
unlink("Rprof.out")

# Look at the log
summary(Log())
}